backend/lambda-requirements.txt
requests==2.31.0
python-jose[cryptography]==3.3.0
boto3==1.34.51
